refactor: LottoGameController로 전달되는 로또 구입 파라미터에 대한 검증과 전달을 DTO 방식으로 로직 위임
